requirements.txt
asgiref==3.2.10
backcall==0.2.0
decorator==4.4.2
Django==3.0.8
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
flake8==3.8.3
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.1
mccabe==0.6.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.2.5
